// apps/marketing/src/generated/OrderDetailsPageLineItemsItemFragment.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: OrderDetailsPageLineItemsItemFragment
// ====================================================

export interface OrderDetailsPageLineItemsItemFragment {
  __typename: "OrderItem";
  id: string;
  title: string;
  quantity: number;
  unitPriceCents: number;
  totalPriceCents: number;
}